requirements.txt
ultralytics
numpy>=1.23.0,<2.0.0
matplotlib>=3.3.0
# opencv-python>=4.6.0
opencv-contrib-python-headless>=4.6.0
pillow>=7.1.2
pyyaml>=5.3.1
requests>=2.23.0
scipy>=1.4.1
torch>=1.8.0
torch>=1.8.0
torchvision>=0.9.0
psutil>=5.8.0
polars>=0.20.0
ultralytics-thop>=2.0.18
python-dotenv